+++ lib/src/lifecycle_event.dart
@@ -22,6 +22,7 @@ class LifecycleEvent {
     this.widgetName,
     this.routeName,
     this.previousRouteName,
+    this.metadata,
   });
 
   final LifecycleEventType type;
@@ -31,6 +32,29 @@ class LifecycleEvent {
   final String? widgetName;
   final String? routeName;
   final String? previousRouteName;
+  final Map<String, Object?>? metadata;
+
+  LifecycleEvent copyWith({
+    LifecycleEventType? type,
+    String? message,
+    DateTime? timestamp,
+    AppLifecycleState? appState,
+    String? widgetName,
+    String? routeName,
+    String? previousRouteName,
+    Map<String, Object?>? metadata,
+  }) {
+    return LifecycleEvent(
+      type: type ?? this.type,
+      message: message ?? this.message,
+      timestamp: timestamp ?? this.timestamp,
+      appState: appState ?? this.appState,
+      widgetName: widgetName ?? this.widgetName,
+      routeName: routeName ?? this.routeName,
+      previousRouteName: previousRouteName ?? this.previousRouteName,
+      metadata: metadata ?? this.metadata,
+    );
+  }
 
   @override
   String toString() => message;